+++ pkg/providers/dns/testdata/zone_mod/expected_zone_mod.tf
@@ -9,11 +9,11 @@ terraform {
 }
 
 locals {
-  zone = "0007770b-08a8-4b5f-a46b-081b772ba605-sbodden-calvin.com"
+  zone = "0007770b-08a8-4b5f-a46b-081b772ba605-test.com"
 }
 
-module "_0007770b-08a8-4b5f-a46b-081b772ba605-sbodden-calvin_com" {
-  source = "modules/_0007770b-08a8-4b5f-a46b-081b772ba605-sbodden-calvin_com"
+module "_0007770b-08a8-4b5f-a46b-081b772ba605-test_com" {
+  source = "modules/_0007770b-08a8-4b5f-a46b-081b772ba605-test_com"
 
   contract = var.contractid
   group    = var.groupid