+++ pkg/providers/papi/create_property_test.go
@@ -1758,6 +1758,56 @@ func TestProcessPolicyTemplates(t *testing.T) {
 			dir:          "basic_with_activation_note",
 			filesToCheck: []string{"property.tf", "variables.tf", "import.sh"},
 		},
+		"property with multiline activation note": {
+			givenData: TFData{
+				Property: TFPropertyData{
+					GroupName:            "test_group",
+					GroupID:              "grp_12345",
+					ContractID:           "test_contract",
+					PropertyResourceName: "test-edgesuite-net",
+					PropertyName:         "test.edgesuite.net",
+					PropertyID:           "prp_12345",
+					ProductID:            "prd_HTTP_Content_Del",
+					ProductName:          "HTTP_Content_Del",
+					RuleFormat:           "latest",
+					IsSecure:             "false",
+					ReadVersion:          "LATEST",
+					EdgeHostnames: map[string]EdgeHostname{
+						"test-edgesuite-net": {
+							EdgeHostname:             "test.edgesuite.net",
+							EdgeHostnameID:           "ehn_2867480",
+							ContractID:               "test_contract",
+							GroupID:                  "grp_12345",
+							ID:                       "",
+							IPv6:                     "IPV6_COMPLIANCE",
+							SecurityType:             "STANDARD-TLS",
+							EdgeHostnameResourceName: "test-edgesuite-net",
+						},
+					},
+					Hostnames: map[string]Hostname{
+						"test.edgesuite.net": {
+							CnameFrom:                "test.edgesuite.net",
+							EdgeHostnameResourceName: "test-edgesuite-net",
+							CertProvisioningType:     "CPS_MANAGED",
+							IsActive:                 true,
+						},
+					},
+					StagingInfo: NetworkInfo{
+						HasActivation:  true,
+						Emails:         []string{"jsmith@akamai.com", "rjohnson@akamai.com"},
+						ActivationNote: "first\nsecond\n\nlast",
+					},
+					ProductionInfo: NetworkInfo{
+						HasActivation:  true,
+						Emails:         []string{"jsmith@akamai.com", "rjohnson@akamai.com"},
+						ActivationNote: "first\nsecond\n",
+					},
+				},
+				Section: "test_section",
+			},
+			dir:          "basic_with_multiline_activation_note",
+			filesToCheck: []string{"property.tf", "variables.tf", "import.sh"},
+		},
 		"property with production activation": {
 			givenData: TFData{
 				Property: TFPropertyData{
@@ -2205,126 +2255,6 @@ func TestTerraformName(t *testing.T) {
 	for name, test := range tests {
 		t.Run(name, func(t *testing.T) {
 			assert.Equal(t, test.expected, TerraformName(test.given))
-		})
-	}
-}
-
-func TestIsMultiline(t *testing.T) {
-	tests := map[string]struct {
-		given    string
-		expected bool
-	}{
-		"has new lines": {
-			given:    "this\nis test\n",
-			expected: true,
-		},
-		"empty": {
-			given:    "",
-			expected: false,
-		},
-		"no new lines": {
-			given:    "no new lines",
-			expected: false,
-		},
-	}
-
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			assert.Equal(t, test.expected, IsMultiline(test.given))
-		})
-	}
-}
-
-func TestNoNewlineAtTheEnd(t *testing.T) {
-	tests := map[string]struct {
-		given    string
-		expected bool
-	}{
-		"has new line at the end": {
-			given:    "this\nis test\n",
-			expected: false,
-		},
-		"has new line in the middle": {
-			given:    "this\nis test",
-			expected: true,
-		},
-		"empty": {
-			given:    "",
-			expected: true,
-		},
-		"no new lines": {
-			given:    "no new lines",
-			expected: true,
-		},
-	}
-
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			assert.Equal(t, test.expected, NoNewlineAtTheEnd(test.given))
-		})
-	}
-}
-
-func TestRemoveLastNewline(t *testing.T) {
-	tests := map[string]struct {
-		given    string
-		expected string
-	}{
-		"has new line at the end": {
-			given:    "this\nis test\n",
-			expected: "this\nis test",
-		},
-		"has new line in the middle": {
-			given:    "this\nis test",
-			expected: "this\nis test",
-		},
-		"empty": {
-			given:    "",
-			expected: "",
-		},
-		"no new lines": {
-			given:    "no new lines",
-			expected: "no new lines",
-		},
-	}
-
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			assert.Equal(t, test.expected, RemoveLastNewline(test.given))
-		})
-	}
-}
-
-func TestGetEOT(t *testing.T) {
-	tests := map[string]struct {
-		given    string
-		expected string
-	}{
-		"has new line": {
-			given:    "this\nis test\n",
-			expected: "EOT",
-		},
-		"has EOT inside": {
-			given:    "this\nEOT",
-			expected: "EOTA",
-		},
-		"empty": {
-			given:    "",
-			expected: "EOT",
-		},
-		"has two EOTs": {
-			given:    "some\nEOT\nEOTA\ntext",
-			expected: "EOTAA",
-		},
-		"has EOT": {
-			given:    "comment\nnewline\nand\nEOT\ninside\n",
-			expected: "EOTA",
-		},
-	}
-
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			assert.Equal(t, test.expected, GetEOT(test.given))
 		})
 	}
 }